Move ValidationRule out of BusinessValidation
