+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Application/Validation/ValidationRule.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace GreenEnergyHub.Charges.Application.Validation.BusinessValidation
+namespace GreenEnergyHub.Charges.Application.Validation
 {
     public class ValidationRule : IValidationRule
     {
